requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
PyJWT==2.10.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
